Create README.md
C# coding exercise:  Console app simulates a Battle Arena where player fights one enemy at a time until the player dies or reaches 10 rounds.  The player is stuck in a dream where the battle continues to loop until the app is closed.  Players start with 10 stamina points and 10 panic points.  Enemies start with 10 hit points.  A random roll of a 6 sided die determines the battle round results.  If an enemy dies, a new one enters the arena and the player continues to fight.